Started working on redoing resource linkage
- Pagination has been reworked to match the specification.
- Handlers now use a recursive merge.
- Changed "href" to "self" under "links".
- Updated tests to match the specification.
- Drop all of the link template handling.

Currently doesn't work on DRF 2 because HyperlinkedIdentityField
doesn't include a queryset, so we can't get the model which is
needed for the type.  This is specifically an issue when handling
the conversion of resources, as we don't have a way to get the
serializer.

Tests are broken in `test_detail` because the parser hasn't been
updated.  Other tests are broken because I haven't started working
on them.

Questions for later:
- I have no idea where to put the old "count" key in pagination.
- I have no idea if/how we should link directly to related objects.
- Should we pass the model in when converting resources?